+++ .idea/dictionaries/Administrator.xml
@@ -1,0 +1,13 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="Administrator">
+    <words>
+      <w>oidc</w>
+      <w>oidcmgr</w>
+      <w>postcss</w>
+      <w>sauthority</w>
+      <w>silentfreshframe</w>
+      <w>silentrefreshframe</w>
+      <w>tripgalleryimplicit</w>
+    </words>
+  </dictionary>
+</component>
