+++ callback.html
@@ -1,0 +1,1 @@
+<!DOCTYPE html><html><head>    <title></title>    <meta charset="utf-8" /></head><body>   <h1>callback;</h1></body></html><!--<script src="https://unpkg.com/oidc-token-manager@1.2.0/dist/oidc-token-manager.min.js"></script>-->
