+++ src/callback.js
@@ -1,0 +1,1 @@
+/** * Created by Administrator on 2017/1/16. */import 'oidc-token-manager';var client_id = 'tripgalleryimplicit';var authority = 'http://10.36.111.213/identity'var post_logout_redirect_uri = window.location.protocol + "//" + window.location.host + "/index.html";var config = {    client_id: client_id,    authority: authority,    //load_user_profile:false}var mgr = new OidcTokenManager(config);mgr.processTokenCallbackAsync().then(function (payload) {    window.location = post_logout_redirect_uri}, function (error) {    alert("Problem Getting Token: " + (error.message || error))})
