+++ src/routes.js
@@ -1,1 +1,1 @@
-/** * Created by Administrator on 2017/1/11. *///域组件import Login from './views/auth/Login.vue'import Home from './views/Home.vue'import Table from './views/user-console/user-mgm.vue'import Form from './views/user-console/add-user-action.vue'import Page3 from './views/user-console/userAdvMgm.vue'import Page4 from './views/test-vuex/click.vue'import Page5 from './views/test-vuex/unimplemented.vue'import echarts from './views/charts/echarts.vue'//嵌套路由表export const routes = [    {        path: '/login',        component: Login,        hidden: true//不显示在导航中    },    {        path: '/',        component: Home,        name: '导航一嵌套',        iconCls: 'el-icon-message',//图标样式class        children: [            //{ path: '/main', component: Main },            {path: '/table', component: Table, name: 'Table'},            {                path: '/form', component: Form, name: 'Form',                iconCls: 'fa fa-bar-chart',                children: [                    {path: '/page4', component: Page4, name: '新增1'},                    {path: '/page5', component: Page5, name: '新增2'}                ]            },            {path: '/page3', component: Page3, name: '页面3'},        ]    },    {        path: '/',        component: Home,        name: '导航二',        iconCls: 'fa fa-id-card-o',        children: [            {path: '/page4', component: Page4, name: '页面4'},            {path: '/page5', component: Page5, name: '页面5'}        ]    },    {        path: '/',        component: Home,        name: '空',        iconCls: 'fa fa-address-card',        leaf: true,//只有一个节点        children: [            {path: '/page6', name: '导航三'}        ]    },    {        path: '/',        component: Home,        name: 'Charts',        iconCls: 'fa fa-bar-chart',        children: [            {path: '/echarts', component: echarts, name: 'echarts'}        ]    }]
+/** * Created by Administrator on 2017/1/11. *///域组件import Login from './views/auth/Login.vue'import Home from './views/Home.vue'import Callback from './views/callback.vue'import Layout from './views/shared/layout.vue'import Table from './views/user-console/user-mgm.vue'import Form from './views/user-console/add-user-action.vue'import Page3 from './views/user-console/userAdvMgm.vue'import Page4 from './views/test-vuex/click.vue'import Page5 from './views/test-vuex/unimplemented.vue'import echarts from './views/charts/echarts.vue'//嵌套路由表export const routes = [    {        path: '/login',        component: Login,        hidden: true//不显示在导航中    },    {        path: '/callback',        component: Callback,        hidden: true//不显示    },    {        path: '/',        component: Layout,        name: '导航一嵌套',        iconCls: 'el-icon-message',//图标样式class        children: [            //{ path: '/main', component: Main },            {path: '/table', component: Table, name: 'Table'},            {                path: '/form', component: Form, name: 'Form',                iconCls: 'fa fa-bar-chart',                children: [                    {path: '/page4', component: Page4, name: '新增1'},                    {path: '/page5', component: Page5, name: '新增2'}                ]            },            {path: '/page3', component: Page3, name: '页面3'},        ]    },    {        path: '/',        component: Home,        name: '导航二',        iconCls: 'fa fa-id-card-o',        children: [            {path: '/page4', component: Page4, name: '页面4'},            {path: '/page5', component: Page5, name: '页面5'}        ]    },    {        path: '/',        component: Home,        name: '空',        iconCls: 'fa fa-address-card',        leaf: true,//只有一个节点        children: [            {path: '/page6', name: '导航三'}        ]    },    {        path: '/',        component: Home,        name: 'Charts',        iconCls: 'fa fa-bar-chart',        children: [            {path: '/echarts', component: echarts, name: 'echarts'}        ]    }]
