+++ src/views/callback.vue
@@ -1,0 +1,1 @@
+<template>    <div>    </div></template><style>    body{        background-color:#ff0000;    }</style><script lang=babel>    import 'oidc-token-manager'    export default{        created: function () {            var config = {                client_id: "tripgalleryimplicit",                authority: "http://10.36.111.213/identity",                //load_user_profile:false            }            var mgr = new OidcTokenManager(config);            mgr.processTokenCallbackAsync().then(function (payload) {                console.log(payload,mgr.access_token)                debugger;                this.$router.replace('/')            }, function (error) {                debugger;                alert("Problem Getting Token: " + (error.message || error))            })        }    }</script>
